I've tried to update the rhythmbox-ampache plugin to be used in Rhythmbox 2.95 (and up) and would like to provide my changes in the attached patch for review and possibly a new branch in this repository.
I based the plugin on HEAD of this repository but used various parts/ideas from the now dysfunct Magnatunes plugin. Major changes are:
- GSettings are used for storing plugin preferences.
- The preferences GUI has been updated. The server url takes now the ampache root and omits the server/xml.server.php extension.
- Refetching is done through the context menu of the entry. The xml file downloaded from the server is used directly as cache file.
- The xml is parsed using sax.

Things I'm not too happy about and will need to be changed:
- rhythmbox entry_delete_by_type() results in an entry->refcount > 0 assertion and Segmentation Fault and therefore do_deactivate() leaves this call out; refetch will cause a Segmentation Fault at this moment because of that
- I would have liked to use Gio for writing the cache file (I use it for downloading the songs information and loading the cache file) but write_async seems to cause garbage at the beginning of each chunk; I'm using python file I/O as a workaround
- I haven't touched the README file

I'm using the plugin myself successfully at the moment.

For information: I've sent in a patch to ampache that should accelerate the xml API songs download to make this plugin more attractive to use.

Any comments are of course very welcome.